+++ changelog.md
@@ -1,5 +1,14 @@
 ## Changelog for TDA API Tool
 
+## [0.2.0]
+### [No-Tracker] - 2020-08-02
+- Moved account management to own file
+- Moved endpoints to own file
+- Updated tdam.py and moved it to top-level
+- Updated readme
+- Moved queries to own file
+- Established decoration for repeatable access check
+
 ## [0.1.0]
 ### [No-Tracker] - 2020-08-01
 - First commit